updated versions
I tried ode23 but unfortunately the outputs were less meaningful so I kept the ode45. I am trying to integrate the immitation rate to the model but unfortunately I loose the oszillationa if I add the imm_rate as in the file no_doleay_G_Model which corresponds to my implementation for the first row in the figure 1 from the paper"The Interplay of Public Intervention and Private Choices in Determining the Outcome of Vaccination Programmes
Alberto d’Onofrio, Piero Manfredi, Piero Poletti". So I commented out that part. In the model since the p_3 is given, instead of using y(3) I changed them to p_3, would you say that is correct?